test: migrate log service tests
